+++ src/main/java/com/example/ecommerce/service/UserService.java
@@ -1,12 +1,19 @@
 package com.example.ecommerce.service;
 
+import com.example.ecommerce.dto.user.CreateUserDto;
+import com.example.ecommerce.dto.user.UserDto;
+import com.example.ecommerce.dto.userDetail.UserDetailDto;
 import com.example.ecommerce.entity.User;
+import com.example.ecommerce.entity.UserDetail;
+import com.example.ecommerce.exception.EntityByGivenNameExistException;
 import com.example.ecommerce.exception.EntityNotFoundException;
 import com.example.ecommerce.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class UserService {
@@ -17,30 +24,54 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public void createUser(User user) {
+    @Transactional
+    public UserDto createUser(CreateUserDto createUserDto) {
+        String username = createUserDto.getUsername();
+
+        if (isExistUser(username)) {
+            throw new EntityByGivenNameExistException(username);
+        }
+        User user = mapToEntity(createUserDto);
         userRepository.save(user);
+        return mapToDto(user);
     }
 
+    @Transactional
     public void deleteUserById(UUID id) {
-        getUser(id);
+        if (!isExistUser(id)) {
+            throw new EntityNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
 
-    public User getUser(UUID id) {
-        return userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id));
+    public UserDto getUser(UUID id) {
+        return mapToDto(userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id)));
     }
 
-    public User getUser(String username) {
-        return userRepository.findUserByUsername(username);
+    public UserDto getUser(String username) {
+        return mapToDto(userRepository.findUserByUsername(username));
     }
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public List<UserDto> getAllUsers() {
+        return userRepository.findAll().stream().map(this::mapToDto).collect(Collectors.toList());
     }
 
-    public void updateUser(User user) {
-        getUser(user.getId());
-        userRepository.save(user);
+    private boolean isExistUser(String username) {
+        return userRepository.existsByUsername(username);
     }
 
+    private boolean isExistUser(UUID id) {
+        return userRepository.existsById(id);
+    }
+
+    private User mapToEntity(CreateUserDto createUserDto) {
+        UUID userId = UUID.randomUUID();
+        UserDetail userDetail = new UserDetail(userId);
+        return new User(userId,createUserDto.getUsername(), createUserDto.getPassword(), userDetail);
+    }
+
+    private UserDto mapToDto(User user) {
+        UserDetailDto userDetailDto = new UserDetailDto(user.getUserDetail().getId(), user.getUserDetail().getFullName());
+        return new UserDto(user.getId(), user.getUsername(), user.getPassword(), userDetailDto);
+    }
 }
